[Gold IV] Title: 기차 여행, Time: 496 ms, Memory: 65220 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/10713. 기차 여행/기차 여행.java
@@ -84,7 +84,7 @@ public class Main {
     private static void calcMinPrice() {
         for (int i = 1; i <= n; i++) {
             int cnt = prefix[i];
-            int price = Math.min(charge[i][0] * cnt, charge[i][2] + charge[i][1] * cnt);
+            long price = Math.min((long)charge[i][0] * cnt, charge[i][2] + (long)charge[i][1] * cnt);
             ans += price;
         }
     }
